project/test: adds more comprehensive tests the resources work as expected
